re organize helper functions
